Правки после первого ревью
