+++ models/user.js
@@ -23,6 +23,7 @@ const userSchema = new mongoose.Schema({
     type: String,
     minlength: 2,
     maxlength: 30,
+    required: true,
     default: 'Иван Иванович',
   },
 });
