+++ package.json
@@ -32,6 +32,7 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "express-winston": "^4.2.0",
+    "helmet": "^4.6.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.0.12",
     "validator": "^13.7.0",
